Rose_Detection/Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
tflite-runtime==2.14.0
httpx==0.27.0
pydantic-settings==2.5.2
opencv-python-headless==4.9.0.80
